refactor: tab.ts => tabs.ts
