+++ src/utils/tabs.ts
@@ -1,10 +1,10 @@
 import browser from 'webextension-polyfill'
 
-import { TAB_MESSAGE } from '~/background/messageListeners/tab'
+import { TABS_MESSAGE } from '~/background/messageListeners/tabs'
 
 export function openLinkInBackground(url: string) {
   return browser.runtime.sendMessage({
-    contentScriptQuery: TAB_MESSAGE.OPEN_LINK_IN_BACKGROUND,
+    contentScriptQuery: TABS_MESSAGE.OPEN_LINK_IN_BACKGROUND,
     url,
   })
 }
